Renaming commands package.
